Remove some unecessary clones
It now passes the control of Snake to the walk method so we don't need
to clone things that will be "discarted", since the old snake is not necessary.

This also fixes the visibility of things, and remove some verboseness of
tests.

Thanks @lsunsi for the great help on explaining me a bit about borrow,
move and visibility in Rust! :D